requirements.txt
redis
PySimpleGUI
pillow
PyMySQL